+++ packages/stylelint-config-itgalaxy/lib/config/rules/block.js
@@ -1,0 +1,16 @@
+/* eslint-disable */
+
+'use strict';
+
+module.exports = {
+    'block-closing-brace-newline-after': 'always',
+    'block-closing-brace-newline-before': 'always',
+    'block-closing-brace-space-after': null,
+    'block-closing-brace-space-before': null,
+    'block-no-empty': true,
+    'block-no-single-line': true,
+    'block-opening-brace-newline-after': 'always-multi-line',
+    'block-opening-brace-newline-before': null,
+    'block-opening-brace-space-after': 'always-single-line',
+    'block-opening-brace-space-before': 'always'
+};