+++ packages/stylelint-config-itgalaxy/lib/config/rules/comment.js
@@ -1,0 +1,23 @@
+/* eslint-disable */
+
+'use strict';
+
+module.exports = {
+    'comment-empty-line-before': [
+        'always',
+        {
+            'except': [
+                'first-nested'
+            ],
+            'ignore': [
+                'stylelint-commands',
+                'between-comments'
+            ]
+        }
+    ],
+    'comment-whitespace-inside': 'always',
+    'comment-word-blacklist': [[
+        '/(T|t)(O|o)(D|d)(O|o)/',
+        '/(F|f)(I|i)(X|x)(M|m)(E|e)/'
+    ]]
+};