+++ packages/stylelint-config-itgalaxy/lib/config/rules/font-family.js
@@ -1,0 +1,7 @@
+/* eslint-disable */
+
+'use strict';
+
+module.exports = {
+    'font-family-name-quotes': 'always-where-recommended'
+};