+++ packages/stylelint-config-itgalaxy/lib/config/rules/font-weight.js
@@ -1,0 +1,14 @@
+/* eslint-disable */
+
+'use strict';
+
+module.exports = {
+    'font-weight-notation': [
+        'numeric',
+        {
+            'ignore': [
+                'relative'
+            ]
+        }
+    ]
+};