+++ packages/stylelint-config-itgalaxy/lib/config/rules/general.js
@@ -1,0 +1,26 @@
+/* eslint-disable */
+
+'use strict';
+
+module.exports = {
+    'indentation': [
+        4,
+        {
+            'indentInsideParens': 'once'
+        }
+    ],
+    'max-empty-lines': 1,
+    'max-line-length': 120,
+    'max-nesting-depth': 7,
+    'no-browser-hacks': null,
+    'no-descending-specificity': null,
+    'no-duplicate-selectors': true,
+    'no-empty-source': true,
+    'no-eol-whitespace': true,
+    'no-extra-semicolons': true,
+    'no-indistinguishable-colors': null,
+    'no-invalid-double-slash-comments': true,
+    'no-missing-eof-newline': true,
+    'no-unknown-animations': true,
+    'no-unsupported-browser-features': null
+};