+++ packages/stylelint-config-itgalaxy/lib/config/rules/keyframe-declaration.js
@@ -1,0 +1,7 @@
+/* eslint-disable */
+
+'use strict';
+
+module.exports = {
+    'keyframe-declaration-no-important': true
+};