+++ packages/stylelint-config-itgalaxy/lib/config/rules/length.js
@@ -1,0 +1,7 @@
+/* eslint-disable */
+
+'use strict';
+
+module.exports = {
+    'length-zero-no-unit': true
+};