+++ packages/stylelint-config-itgalaxy/lib/config/rules/media-query.js
@@ -1,0 +1,7 @@
+/* eslint-disable */
+
+'use strict';
+
+module.exports = {
+    'media-query-parentheses-space-inside': 'never'
+};