+++ packages/stylelint-config-itgalaxy/lib/config/rules/number.js
@@ -1,0 +1,9 @@
+/* eslint-disable */
+
+'use strict';
+
+module.exports = {
+    'number-leading-zero': 'always',
+    'number-max-precision': 8,
+    'number-no-trailing-zeros': true
+};