+++ packages/stylelint-config-itgalaxy/lib/config/rules/property.js
@@ -1,0 +1,10 @@
+/* eslint-disable */
+
+'use strict';
+
+module.exports = {
+    'property-blacklist': null,
+    'property-case': 'lower',
+    'property-no-vendor-prefix': true,
+    'property-whitelist': null
+};