+++ packages/stylelint-config-itgalaxy/lib/config/rules/root-rule.js
@@ -1,0 +1,7 @@
+/* eslint-disable */
+
+'use strict';
+
+module.exports = {
+    'root-no-standard-properties': true
+};