+++ packages/stylelint-config-itgalaxy/lib/config/rules/selector-list.js
@@ -1,0 +1,10 @@
+/* eslint-disable */
+
+'use strict';
+
+module.exports = {
+    'selector-list-comma-newline-after': 'always',
+    'selector-list-comma-newline-before': null,
+    'selector-list-comma-space-after': null,
+    'selector-list-comma-space-before': 'never'
+};