+++ packages/stylelint-config-itgalaxy/lib/config/rules/selector.js
@@ -1,0 +1,48 @@
+/* eslint-disable */
+
+'use strict';
+
+module.exports = {
+    'selector-attribute-brackets-space-inside': 'never',
+    'selector-attribute-operator-blacklist': null,
+    'selector-attribute-operator-space-after': 'never',
+    'selector-attribute-operator-space-before': 'never',
+    'selector-attribute-operator-whitelist': null,
+    'selector-class-pattern': '^[a-z][a-z0-9-_]{0,44}[a-z0-9]$',
+    'selector-combinator-space-after': 'always',
+    'selector-combinator-space-before': 'always',
+    'selector-id-pattern': '^[a-z][a-z0-9-_]{0,30}[a-z0-9]$',
+    'selector-max-compound-selectors': 5,
+    'selector-max-specificity': '0,5,0',
+    'selector-no-attribute': null,
+    'selector-no-combinator': null,
+    'selector-no-id': null,
+    'selector-no-qualifying-type': [
+        true,
+        {
+            'ignore': [
+                'class'
+            ]
+        }
+    ],
+    'selector-no-type': null,
+    'selector-no-universal': true,
+    'selector-no-vendor-prefix': true,
+    'selector-pseudo-class-case': 'lower',
+    'selector-pseudo-class-no-unknown': true,
+    'selector-pseudo-class-parentheses-space-inside': 'never',
+    'selector-pseudo-element-case': 'lower',
+    'selector-pseudo-element-colon-notation': 'double',
+    'selector-pseudo-element-no-unknown': true,
+    'selector-root-no-composition': true,
+    'selector-type-case': 'lower',
+    'selector-type-no-unknown': [
+        true,
+        {
+            'ignoreTypes': [
+                'noindex'
+            ]
+        }
+    ],
+    'selector-max-empty-lines': 0
+};