+++ packages/stylelint-config-itgalaxy/lib/config/rules/shorthand-property.js
@@ -1,0 +1,7 @@
+/* eslint-disable */
+
+'use strict';
+
+module.exports = {
+    'shorthand-property-no-redundant-values': true
+};