+++ packages/stylelint-config-itgalaxy/lib/config/rules/string.js
@@ -1,0 +1,8 @@
+/* eslint-disable */
+
+'use strict';
+
+module.exports = {
+    'string-no-newline': true,
+    'string-quotes': 'double'
+};