+++ packages/stylelint-config-itgalaxy/lib/config/rules/stylelint-disable-comment.js
@@ -1,0 +1,7 @@
+/* eslint-disable */
+
+'use strict';
+
+module.exports = {
+    'stylelint-disable-reason': 'always-after'
+};