+++ packages/stylelint-config-itgalaxy/lib/config/rules/unit.js
@@ -1,0 +1,23 @@
+/* eslint-disable */
+
+'use strict';
+
+module.exports = {
+    'unit-whitelist': null,
+    'unit-case': 'lower',
+    'unit-no-unknown': true,
+    'unit-blacklist': [
+        'ex',
+        'cm',
+        'mm',
+        'q',
+        'in',
+        'pt',
+        'pc',
+        'ch',
+        'vm',
+        'grad',
+        'turn',
+        'rad'
+    ]
+};