+++ packages/stylelint-config-itgalaxy/lib/config/rules/value.js
@@ -1,0 +1,8 @@
+/* eslint-disable */
+
+'use strict';
+
+module.exports = {
+    'value-no-vendor-prefix': true,
+    'value-keyword-case': 'lower'
+};